Reformatting
I have reformatted certain parts of the document.

I have tried to consistenly use `` to Markdown text that should be shown as code (such as print() ).

Also, I harmonized the font color in Markdown to match the color used in the Python code in Jupyter Notebook.

Other fixes are a few rephrasings and typos.