More reformatting and minor clarifications
I have found a few additional typos and corrected them

In the topic data type conversion, I have added a negative example for a conversion that does not work.

For boolean operators I also added a negative example, to illustrate when they do not work.

I have added Roman numerals to the first two topics (Introduction and "How to use the Jupyter Notebook"). The conclusion is not numbered.